refactor: 안드로이드 프로젝트 생성으로 인한 폴더 구조 변경
