+++ domain/src/main/kotlin/Application.kt
@@ -1,0 +1,9 @@
+import controller.GameController
+
+fun main() {
+    Thread.setDefaultUncaughtExceptionHandler { _, exception ->
+        exception.printStackTrace()
+    }
+    val game = GameController()
+    game.process()
+}
