+++ domain/src/main/kotlin/controller/Mapper.kt
@@ -1,0 +1,6 @@
+package controller
+
+interface Mapper<S, D> {
+    fun domainToDTO(source: S): D
+    fun dtoToDomain(source: D): S
+}
