+++ domain/src/main/kotlin/domain/Color.kt
@@ -1,0 +1,8 @@
+package domain
+
+enum class Color {
+    BLACK,
+    WHITE;
+
+    fun turnColor(): Color = if (this == BLACK) WHITE else BLACK
+}
