+++ domain/src/main/kotlin/domain/CoordinateReader.kt
@@ -1,0 +1,5 @@
+package domain
+
+fun interface CoordinateReader {
+    fun read(color: Color): Coordinate
+}
