+++ domain/src/main/kotlin/domain/Stone.kt
@@ -1,0 +1,3 @@
+package domain
+
+data class Stone(val color: Color, val coordinate: Coordinate)
