+++ domain/src/main/kotlin/dto/ColorDTO.kt
@@ -1,0 +1,6 @@
+package dto
+
+enum class ColorDTO {
+    BLACK,
+    WHITE
+}
