+++ domain/src/main/kotlin/dto/PointDTO.kt
@@ -1,0 +1,6 @@
+package dto
+
+data class PointDTO(
+    val x: Int,
+    val y: Int
+)
