+++ domain/src/main/kotlin/dto/StoneDTO.kt
@@ -1,0 +1,6 @@
+package dto
+
+data class StoneDTO(
+    val color: ColorDTO,
+    val coordinate: PointDTO,
+)
