+++ domain/src/main/kotlin/view/RenderBoard.kt
@@ -1,0 +1,8 @@
+package view
+
+import dto.PointDTO
+import dto.StoneDTO
+
+interface RenderBoard {
+    fun render(stones: List<StoneDTO>, size: PointDTO): String
+}
