--- src/main/kotlin/controller/Mapper.kt
@@ -1,6 +1,0 @@
-package controller
-
-interface Mapper<S, D> {
-    fun domainToDTO(source: S): D
-    fun dtoToDomain(source: D): S
-}
