--- src/main/kotlin/domain/CoordinateReader.kt
@@ -1,5 +1,0 @@
-package domain
-
-fun interface CoordinateReader {
-    fun read(color: Color): Coordinate
-}
