--- src/main/kotlin/domain/Stone.kt
@@ -1,3 +1,0 @@
-package domain
-
-data class Stone(val color: Color, val coordinate: Coordinate)
