--- src/main/kotlin/dto/ColorDTO.kt
@@ -1,6 +1,0 @@
-package dto
-
-enum class ColorDTO {
-    BLACK,
-    WHITE
-}
