--- src/main/kotlin/dto/PointDTO.kt
@@ -1,6 +1,0 @@
-package dto
-
-data class PointDTO(
-    val x: Int,
-    val y: Int
-)
