--- src/main/kotlin/dto/StoneDTO.kt
@@ -1,6 +1,0 @@
-package dto
-
-data class StoneDTO(
-    val color: ColorDTO,
-    val coordinate: PointDTO,
-)
