--- src/test/kotlin/domain/StoneForm.kt
@@ -1,9 +1,0 @@
-package domain
-
-fun BlackStone(x: Int, y: Int): Stone {
-    return Stone(Color.BLACK, Coordinate.from(x, y)!!)
-}
-
-fun WhiteStone(x: Int, y: Int): Stone {
-    return Stone(Color.WHITE, Coordinate.from(x, y)!!)
-}
